提高代码可读性

+++ src/chap11_gaussian_mixture/GMM.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 from typing import Tuple, List 
 
 # 生成混合高斯分布数据
-def generate_data(n_samples=1000):
+def generate_data(n_samples = 1000):
     """生成混合高斯分布数据集
     
     Args:
@@ -75,7 +75,7 @@ def generate_data(n_samples=1000):
     return X[shuffle_idx], y_true[shuffle_idx]
 
 # 自定义logsumexp函数
-def logsumexp(log_p, axis=1, keepdims=False):
+def logsumexp(log_p, axis = 1, keepdims = False):
     """优化后的logsumexp实现，包含数值稳定性增强和特殊case处理
     
     计算log(sum(exp(log_p)))，通过减去最大值避免数值溢出
@@ -93,12 +93,12 @@ def logsumexp(log_p, axis=1, keepdims=False):
     
     # 处理空输入情况
     if log_p.size == 0:                                                     # 检查输入的对数概率数组是否为空
-        return np.array(-np.inf, dtype=log_p.dtype)                         # 返回与输入相同数据类型的负无穷值
+        return np.array(-np.inf, dtype = log_p.dtype)                         # 返回与输入相同数据类型的负无穷值
     
     # 计算最大值（处理全-inf输入）
-    max_val = np.max(log_p, axis=axis, keepdims=True)                       # 计算沿指定轴的最大值
+    max_val = np.max(log_p, axis = axis, keepdims = True)                       # 计算沿指定轴的最大值
     if np.all(np.isneginf(max_val)):                                        # 检查是否所有最大值都是负无穷
-        return max_val.copy() if keepdims else max_val.squeeze(axis=axis)   # 根据keepdims返回适当形式
+        return max_val.copy() if keepdims else max_val.squeeze(axis = axis)   # 根据keepdims返回适当形式
     
     # 计算修正后的指数和（处理-inf输入）
     # 安全计算指数和：先减去最大值，再计算指数
@@ -110,7 +110,7 @@ def logsumexp(log_p, axis=1, keepdims=False):
     
     # 处理全-inf输入的特殊case
     if np.any(np.isneginf(log_p)) and not np.any(np.isfinite(log_p)):       # 判断是否所有有效值都是-inf
-        result = max_val.copy() if keepdims else max_val.squeeze(axis=axis) # 根据keepdims参数的值返回max_val的适当形式
+        result = max_val.copy() if keepdims else max_val.squeeze(axis = axis) # 根据keepdims参数的值返回max_val的适当形式
     return result                                                           # 返回处理后的结果，保持与正常情况相同的接口
 
 # 高斯混合模型类
@@ -149,7 +149,7 @@ class GaussianMixtureModel:
         self.pi = np.ones(self.n_components) / self.n_components
         
         # 随机选择样本点作为初始均值
-        self.mu = X[np.random.choice(n_samples, self.n_components, replace=False)]
+        self.mu = X[np.random.choice(n_samples, self.n_components, replace = False)]
         
         # 初始化协方差矩阵为单位矩阵
         self.sigma = np.array([np.eye(n_features) for _ in range(self.n_components)])
@@ -167,7 +167,7 @@ class GaussianMixtureModel:
                 log_prob[:, k] = np.log(self.pi[k]) + self._log_gaussian(X, self.mu[k], self.sigma[k]) # 计算第k个高斯混合成分的对数概率密度，并存储在log_prob的第k列
             
             # 使用logsumexp计算归一化因子，确保数值稳定性
-            log_prob_sum = logsumexp(log_prob, axis=1, keepdims=True)
+            log_prob_sum = logsumexp(log_prob, axis = 1, keepdims = True)
             
             # 计算后验概率（responsibility）：gamma_{ik} = P(z_i=k|x_i)
             gamma = np.exp(log_prob - log_prob_sum)
